requirements.txt
PyQt5>=5.15.0
onnxruntime>=1.12.0
Pillow>=8.0.0
numpy>=1.21.0 
opencv-python>=4.0.0
